Refactor fields.jl
Code for working with fields is now refactored and documented. A slight
modification is done how to define variable fields. Earlier the syntax
was to define a vector with some content, but this syntax was
not unambiguous, e.g. `update!(el, "f", [1.0, 2.0])` could mean either
constant vector field or variable scalar field in case of Seg2. Now
variable field is defined using Tuple so it's always clear what kind of
field is defined.