Change of strategy
It's better to implement three different kind of DOFMaps. This is the 
first one and should be compatible with the curren definition. We can 
then later on extend to FixedDOFMap where we can effectively make a 
bandwidth-minimizing permutation ourself and later on VariableDOFMap so 
that each node can have different number of dofs.